+++ provider/pkg/openapi/match.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	jsonpatch "github.com/evanphx/json-patch"
+	jsonpatch "github.com/pulumi/json-patch"
 	"k8s.io/client-go/util/jsonpath"
 )